+++ packages/devtools_app/lib/src/flutter/scaffold.dart
@@ -2,7 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:async';
 import 'dart:math' as math;
 
 import 'package:flutter/foundation.dart';
@@ -82,9 +81,7 @@ class DevToolsScaffoldState extends State<DevToolsScaffold>
   /// coordinate their animation when the tab selection changes.
   TabController _tabController;
 
-  Screen _currentScreen;
-  final StreamController<Screen> _screenController =
-      StreamController.broadcast();
+  ValueNotifier<Screen> _currentScreen;
 
   ImportController _importController;
 
@@ -150,17 +147,10 @@ class DevToolsScaffoldState extends State<DevToolsScaffold>
     _tabController?.dispose();
     _tabController = TabController(length: widget.tabs.length, vsync: this);
 
-    _handleScreenChange(widget.tabs[_tabController.index]);
+    _currentScreen = ValueNotifier(widget.tabs[_tabController.index]);
     _tabController.addListener(() {
-      _handleScreenChange(widget.tabs[_tabController.index]);
+      _currentScreen.value = widget.tabs[_tabController.index];
     });
-  }
-
-  void _handleScreenChange(Screen screen) {
-    if (screen?.type != _currentScreen?.type) {
-      _currentScreen = screen;
-      _screenController.add(_currentScreen);
-    }
   }
 
   /// Pushes tab changes into the navigation history.
@@ -209,11 +199,8 @@ class DevToolsScaffoldState extends State<DevToolsScaffold>
         ),
     ];
 
-    return StreamProvider<Screen>(
-      create: (BuildContext context) {
-        return _screenController.stream;
-      },
-      initialData: _currentScreen,
+    return ValueListenableProvider<Screen>.value(
+      value: _currentScreen,
       child: DragAndDrop(
         handleDrop: _importController.importData,
         child: AnimatedBuilder(
